+++ docs/source/changes.rst
@@ -10,6 +10,7 @@ Release 0.5.2 (under development)
 * Add support for using either CUSOLVER or CULA for computing SVD.
 * Compressed Dynamic Mode Decomposition (enh. by N. Benjamin Erichson).
 * Wrappers for CUFFT size estimation (enh. by Luke Pfister).
+* Wrappers for CUBLAS-XT functions.
 
 Release 0.5.1 - (October 30, 2015)
 ----------------------------------